+++ packages/coin-support-canton/src/operations/prepareTransaction/index.ts
@@ -43,20 +43,24 @@ const validateAddresses = async (
   return outputAddressValidation;
 };
 
-const selectUtxos = (targetAmount: string, utxos: any[]): string[] => {
+const selectUtxos = (targetAmount: string, sortedUtxos: any[]): string[] => {
   // Select utxos whose total sum is greater than or equal to the targetAmount
-  // We need to select the maximum number of utxos that can be used to send the targetAmount
-  // so that number of utxos are reduced to minimum
-  // But this might increase the traffic cost as we are sending more utxos
-  // So we need to select the utxos in a way that the traffic cost is minimized:
-  // We will implement a better strategy later
+  // We need to select the minimum number of utxos that can be used to send the targetAmount
+  // So that the traffic cost is minimized
+  // and transaction is faster
 
-  // Sort the utxos by value in ascending order
-  const sortedUtxos = utxos.sort((a, b) =>
-    new BigNumber(a.interfaceViewValue.amount)
-      .minus(b.interfaceViewValue.amount)
-      .toNumber(),
-  );
+  // considering the utxos in descending order of value
+  // Single UTXO >= targetAmount
+  for (let i = sortedUtxos.length - 1; i >= 0; i -= 1) {
+    const currentUtxo = sortedUtxos[i];
+    if (
+      new BigNumber(
+        currentUtxo.interfaceViewValue.amount,
+      ).isGreaterThanOrEqualTo(targetAmount)
+    ) {
+      return [currentUtxo.contractId];
+    }
+  }
 
   let totalAmount = new BigNumber(0);
   const selectedUtxos: string[] = [];
@@ -69,21 +73,7 @@ const selectUtxos = (targetAmount: string, utxos: any[]): string[] => {
     }
   }
 
-  // Remove utxos from the start of the array if the total amount is greater than the target amount
-  let i = 0;
-  while (i < selectedUtxos.length) {
-    const currentUtxoAmount = sortedUtxos[i].interfaceViewValue.amount;
-    if (
-      totalAmount.minus(currentUtxoAmount).isGreaterThanOrEqualTo(targetAmount)
-    ) {
-      totalAmount = totalAmount.minus(currentUtxoAmount);
-      i += 1;
-    } else {
-      break;
-    }
-  }
-
-  return selectedUtxos.slice(i);
+  return selectedUtxos;
 };
 
 export const prepareTransaction = async (
@@ -141,12 +131,13 @@ export const prepareTransaction = async (
     txn.userInputs.outputs[0].amount = output.amount;
   };
 
-  let hasEnoughBalance: boolean;
-
   if (txn.userInputs.isSendAll) {
     calculateMaxSend();
   }
 
+  const isValidAmount = !sendAmount.isNaN() && !sendAmount.isZero();
+
+  let hasEnoughBalance: boolean;
   hasEnoughBalance =
     sendAmount.isNaN() ||
     new BigNumber(account.balance).isGreaterThanOrEqualTo(
@@ -167,8 +158,6 @@ export const prepareTransaction = async (
     }
   }
 
-  const isValidAmount = !sendAmount.isNaN() && !sendAmount.isZero();
-
   const isValidInputs =
     output.address?.length &&
     isValidAmount &&
@@ -187,7 +176,7 @@ export const prepareTransaction = async (
         instrument,
         memo: output.memo,
         expiryDate: output.expiryDate,
-        inputUtxos: selectUtxos(output.amount, txn.staticData.utxos),
+        inputUtxos: selectUtxos(output.amount, txn.staticData.sortedUtxos),
       },
       keyDB,
     );