+++ packages/cysync-core-constants/src/i18n/types.ts
@@ -116,6 +116,7 @@ interface LangButtons {
   mergeDelegation: string;
   accept: string;
   earnRewards: string;
+  proceed: string;
 }
 
 interface LangDashboard {
@@ -1433,6 +1434,14 @@ interface LangDialogs {
           cancel: string;
         };
       };
+    };
+    login: {
+      success: {
+        title: string;
+      };
+    };
+    syncAccountPrompt: {
+      title: string;
     };
   };
 }