+++ packages/ui/src/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './useAccordion';
 export * from './addKeyboardEvents';
+export * from './useOverflow';