+++ packages/coin-support-solana/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cypherock/coin-support-solana
 
+## 0.2.8-pre.0
+
+### Patch Changes
+
+- Updated dependencies [[`223a00aa`](https://github.com/Cypherock/cypherock-cysync/commit/223a00aa4cc8b06f75b75b58cbebb7dfcbddeb76)]:
+  - @cypherock/coins@0.7.0-pre.0
+  - @cypherock/coin-support-utils@0.1.8-pre.0
+
 ## 0.2.7
 
 ### Patch Changes